stock-ai-backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
requests
bcrypt
pyjwt
redis
python-multipart
